Allow unlocked USB data access (1/2)
by @KreAch3R

This bypasses the marshmallow "regression" of blocking and resetting USB data access
after disconnecting. It also allows saving the default configuration and it survives a reboot.
Originally, this "feature" was added by Google for security reasons
so it should be advised to use the below setting under caution.

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>
Signed-off-by: sub77 <sub77@ymail.com>

Conflicts:
	services/usb/java/com/android/server/usb/UsbDeviceManager.java

Change-Id: I35a0de02e7e25d88d83041bcfac33306fe787590
Signed-off-by: sub77 <sub77@ymail.com>